LIBFCREPO-496. Minor fixes to umd-fcrepo-broadcast
This fixes are unrelated to LIBFCREPO-496, but I noticed them when
working on the issue, and they didn't seem wortwhile to create a
separate issue for.

The "log" variable is not used by BroadcastDispatched (and certainly
shouldn't be part of the "toString" output, so it was removed.

The "<type>jar</type>" is unnecessary in the pom.xml file, so it was
removed.

https://issues.umd.edu/browse/LIBFCREPO-496